+++ GW2EIBuilders/JsonModels/JsonActors/JsonActor.cs
@@ -115,14 +115,14 @@ namespace GW2EIBuilders.JsonModels
         public List<int>[] Damage1S { get; internal set; }
         [JsonProperty]
         /// <summary>
-        /// Array of int representing 1S breakbar damage points \n
+        /// Array of double representing 1S breakbar damage points \n
         /// Length == # of phases
         /// </summary>
         /// <remarks>
         /// If the duration of the phase in seconds is non integer, the last point of this array will correspond to the last point  \n
         /// ex: duration === 15250ms, the array will have 17 elements [0, 1000,...,15000,15250]
         /// </remarks>
-        public List<int>[] BreakbarDamage1S { get; internal set; }
+        public List<double>[] BreakbarDamage1S { get; internal set; }
         [JsonProperty]
         /// <summary>
         /// Array of int[2] that represents the number of conditions \n
@@ -191,7 +191,7 @@ namespace GW2EIBuilders.JsonModels
             if (settings.RawFormatTimelineArrays)
             {
                 Damage1S = new List<int>[phases.Count];
-                BreakbarDamage1S = new List<int>[phases.Count];
+                BreakbarDamage1S = new List<double>[phases.Count];
                 for (int i = 0; i < phases.Count; i++)
                 {
                     Damage1S[i] = actor.Get1SDamageList(log, i, phases[i], null);
